+++ HISTORY.md
@@ -1,9 +1,10 @@
 # Version History
 
-## v1.0.1 (unreleased)
+## v1.1.0 (unreleased)
 
 - Added `"values"."sourcesColumnWidth"` field to template file
 - Increased default width of column 1 to 285
+- Added all editors of spreadsheet (except volunteer) to protected range
 
 ## v1.0.0 (2023-03-03)